Test: wrote tests for the Output component
